matcher/requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
Faker==13.14.0
mysql-connector==2.2.9
mysql-connector-python==8.0.29
protobuf==4.21.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
